lab2 added (продолжить)
