+++ dmitriYakovlev/components/routers/1.txt
@@ -1,0 +1,3 @@
+Название: S.T.A.L.K.E.R. 2: Heart Of Chernobyl
+Дата выхода: 28/04/2022
+Жанр: Экшен, Шутер, Вид от первого лица, Постапкалипитика
