+++ dmitriYakovlev/components/routers/2.txt
@@ -1,0 +1,3 @@
+Название: Корсары: Черная метка
+Дата выхода: 2022
+Жанр: Ролевая игра, Симулятор, Экшен, Вид от третьего лица, Историческая
