+++ dmitriYakovlev/components/routers/3.txt
@@ -1,0 +1,3 @@
+Название: Dying Light 2 Stay Human
+Дата выхода: 4 февраля 2022
+Жанр: Экшен, Шутер, Хоррор, Вид от первого лица, Постапокалиптика
