+++ dmitriYakovlev/components/routers/4.txt
@@ -1,0 +1,3 @@
+Название: Metro 4
+Дата выхода: 2022
+Жанр: Экшен, Шутер, Вид от первого лица, Постапокалиптика
